pyproject.toml
[project]
name = "rag-and-vectordbs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "langchain-community>=0.3.8",
    "langchain>=0.3.9",
    "langchain-openai>=0.2.10",
    "langchain-pinecone>=0.2.0",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.0.1",
]
